tools/pyproject.toml
[project]
name = "startup-game-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.63.0",
    "openai>=2.20.0",
    "pillow>=11.0.0",
    "python-dotenv==1.0.1",
    "requests>=2.32.5",
]

[project.scripts]
process-sprites = "process_sprites:main"
slice-building = "slice_building:main"
